requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.9
cloudinary==1.28.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0
django-cloudinary-storage==0.3.0
django-environ==0.8.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
environ==1.0
filelock==3.4.0
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.3
pygame==2.1.0
pygame-gui==0.5.7
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.10.0
whitenoise==5.3.0
